+++ checkpoint/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - #v1 Add `use_load_and_broadcast` option.
+- Add PyTorch DCP (Distributed Checkpoint) to the benchmark suite.
 
 ### Removed